refactor: UserService가 TermAgreementRepository가 아니라 TermAgreementService에 의존하도록 수정
