+++ mitube-app/src/main/java/com/misim/controller/model/TermResponseDto.java
@@ -1,10 +1,7 @@
 package com.misim.controller.model;
 
-import com.misim.entity.Term;
-import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
-import lombok.RequiredArgsConstructor;
 
 @Getter
 @Builder
